Split Block into separate modules
